+++ include/metaball.hh
@@ -4,13 +4,13 @@
 class Metaball : public Blob
 {
 public:
-    Metaball(Vector3 lower_left_corner, float side_length, int step,
-             float threshold)
+    Metaball(Vector3 lower_left_corner, double side_length, int step,
+             double threshold)
         : Blob(lower_left_corner, side_length, step, threshold)
     {}
 
     std::vector<Sphere> balls;
 
 private:
-    float evaluate_potential(Vector3 point);
+    double evaluate_potential(Vector3 point);
 };
